+++ Sources/VComponents/Components/Modals/Alert/VAlertUIModel.swift
@@ -156,7 +156,7 @@ public struct VAlertUIModel {
         public var shadowRadius: CGFloat = 0
         
         /// Shadow offset. Set to `zero`.
-        public var shadowOffset: CGSize = .zero
+        public var shadowOffset: CGPoint = .zero
         
         /// Dimming view color.
         public var dimmingView: Color = GlobalUIModel.Common.dimmingViewColor