+++ Sources/VComponents/Components/Modals/Bottom Sheet/VBottomSheet.swift
@@ -103,8 +103,7 @@ struct VBottomSheet<Content>: View
                 .shadow(
                     color: uiModel.colors.shadow,
                     radius: uiModel.colors.shadowRadius,
-                    x: uiModel.colors.shadowOffset.width,
-                    y: uiModel.colors.shadowOffset.height
+                    offset: uiModel.colors.shadowOffset
                 )
                     
             VStack(spacing: 0, content: {