+++ Sources/VComponents/Components/Modals/Modal/VModal.swift
@@ -94,8 +94,7 @@ struct VModal<Content>: View
             .shadow(
                 color: uiModel.colors.shadow,
                 radius: uiModel.colors.shadowRadius,
-                x: uiModel.colors.shadowOffset.width,
-                y: uiModel.colors.shadowOffset.height
+                offset: uiModel.colors.shadowOffset
             )
     }