+++ Sources/VComponents/Components/Modals/SideBar/VSideBar.swift
@@ -69,8 +69,7 @@ struct VSideBar<Content>: View where Content: View {
                 .shadow(
                     color: uiModel.colors.shadow,
                     radius: uiModel.colors.shadowRadius,
-                    x: uiModel.colors.shadowOffset.width,
-                    y: uiModel.colors.shadowOffset.height
+                    offset: uiModel.colors.shadowOffset
                 )
 
             content()